DRTII-1691 Tests around queue export function
